+++ Assets/_Code/_Tests/GamePlay/DamageSystem/DamagerTakerTest.cs
@@ -13,8 +13,8 @@ public class DamagerTakerTest
         GameObject testObject = new GameObject("DamageTakerGameObject");
         testObject.AddComponent<DamageTaker>();
         DamageTaker damageTaker = testObject.GetComponent<DamageTaker>();
-        double initialHPBeforeDamage = damageTaker.GetCurrentDamage();
-        double damage = 10D;
+        float initialHPBeforeDamage = damageTaker.GetCurrentDamage();
+        float damage = 10F;
         DamagePayload damagePayload = new DamagePayload(damage, DamageType.Blunt);
         damageTaker.TakeDamage(damagePayload);
         Assert.That(damageTaker.GetCurrentDamage(), Is.EqualTo(initialHPBeforeDamage));
